added shuffled indices cache tests
